Refactored graph component files names to be more readable
