+++ frontend/src/components/tests/orgs-list.test.js
@@ -1,11 +1,11 @@
-// import { render, screen } from '@testing-library/react';
-// import { BrowserRouter as Router } from 'react-router-dom';
-// import OrgsList from '../orgs-list'
+import { render, screen } from '@testing-library/react';
+import { BrowserRouter as Router } from 'react-router-dom';
+import OrgsList from '../orgs-list'
 
-// test('Renders table with headings', () => {
-//   render(<Router>
-//     <OrgsList/>
-//   </Router>);
-//   const tableheader = screen.getByText(/Logs/i);
-//   expect(tableheader).toBeInTheDocument();
-// });
+test('Renders table with headings', () => {
+  render(<Router>
+    <OrgsList/>
+  </Router>);
+  const tableheader = screen.getByTestId('OrgsList')
+  expect(tableheader).toBeInTheDocument();
+});
